[Fist Commit] Structure of the project.
The application will start with a slpash,
and will use a worker to retrieve the data
from http://site.api.espn.com/apis/site/v2/sports/basketball/nba/
after retrieved it, will insert all the information
(in this case is a list of teams with they information) into a database.
After the worker is done the StandingFragment will show
the standing for the teams.

Signed-off-by: Ignacio Sambataro <ignaciosamba@gmail.com>